third_party/WebKit/LayoutTests/fast/css/webkit-mask-crash-implicit-expected.txt

FAIL document.styleSheets[1].rules[0].style.cssText should be -webkit-mask-repeat-x: repeat; -webkit-mask-repeat-y: inherit;. Was -webkit-mask-repeat: inherit;.
PASS document.styleSheets[1].rules[0].style.getPropertyValue('-webkit-mask') is ""
FAIL document.styleSheets[1].rules[0].style.getPropertyValue('-webkit-mask-repeat') should be . Was inherit.
PASS successfullyParsed is true

TEST COMPLETE
